Check and raise exception for partial responses
We do not want to pass along partial responses to the JSON parser
as this usually results in JSON parser errors due to invalid data.
This adds checking the return_code for :partial_file which indicates
from ethon/libcurl that the Content-Length header does not match the
received response.
Raise an exception if that happens.

We are ignoring other return_codes for now as we do not have the need
to check for them.